move install path generation to util function
